test(e2e): poll for the offline transition instead of sleeping
`reconnect()` tears the socket down in a background task, so the fixed
100ms sleep that followed it was the only thing standing between an
offline test and a client that had not gone offline yet. Add
`TestClient::wait_for_disconnected()` and wait on the connection flag,
which is the observable here since `Event::Disconnected` is suppressed
for expected disconnects. Sleeps that followed `disconnect()` are dropped
outright: it already awaits the run task.

Also hoist the helpers that had been copied per test file (`has_child`,
`scan_sessions` and its address builder) into the shared crate, keeping
the single form that reports both the JID and whether a session record
exists.
